Let modules define and register their custom template filters
