+++ flirror/modules/stocks/__init__.py
@@ -22,6 +22,11 @@ def get():
     )
 
 
+@stocks_module.app_template_filter()
+def list_filter(list_of_dicts_to_filter, key):
+    return [d[key] for d in list_of_dicts_to_filter]
+
+
 @stocks_module.crawler("stocks-crawler")
 def crawl(crawler_id, app, api_key, symbols, mode="table"):
     ts = TimeSeries(key="YOUR_API_KEY")
