--- flirror/templates/modules/map.html
@@ -1,11 +1,0 @@
-{% extends "module.html" %}
-
-{% block body %}
-<div class="card-body">
-    <div class="text-right"><small><i class="fas fa-sync-alt"></i> {{ module.data._timestamp | prettydate }}</small></div>
-    <p class="card-text">
-        <i class="fas fa-map"></i>
-        This could show a map for the route from {{ module.data.route.from }} to {{ module.data.route.to }}
-    </p>
-</div>
-{% endblock %}
