+++ flirror/utils.py
@@ -7,31 +7,6 @@ import arrow
 
 LOGGER = logging.getLogger(__name__)
 
-weather_icons = {
-    "01d": "wi wi-day-sunny",
-    "02d": "wi wi-day-cloudy",
-    "03d": "wi wi-cloud",
-    "04d": "wi wi-cloudy",
-    "09d": "wi wi-day-showers",
-    "10d": "wi wi-day-rain",
-    "11d": "wi wi-day-thunderstorm",
-    "13d": "wi wi-day-snow",
-    "50d": "wi wi-dust",
-    "01n": "wi wi-night-clear",
-    "02n": "wi wi-night-alt-cloudy",
-    "03n": "wi wi-cloud",
-    "04n": "wi wi-cloudy",
-    "09n": "wi wi-night-showers-rain",
-    "10n": "wi wi-night-rain",
-    "11n": "wi wi-night-thunderstorm",
-    "13n": "wi wi-night-snow",
-    "50n": "wi wi-dust",
-}
-
-
-def weather_icon(icon_name):
-    return weather_icons.get(icon_name)
-
 
 def prettydate(date):
     """
@@ -73,10 +48,6 @@ def parse_interval_string(interval_string):
 def format_time(timestamp, format):
     date = datetime.utcfromtimestamp(timestamp)
     return date.strftime(format)
-
-
-def list_filter(list_of_dicts_to_filter, key):
-    return [d[key] for d in list_of_dicts_to_filter]
 
 
 def clean_string(string):
